requirements.txt
# --- Production Dependencies ---
openreview-py # Pin your version for reproducibility
requests>=2.28
# typing is usually included in modern Python, but listing is safe for older envs
typing_extensions>=4.0 

# --- Development/CI Dependencies ---
# These are needed by the GitHub Actions setup:
ruff>=0.1.0
black>=23.0
pytest>=7.0
pytest-cov>=4.0
python-dotenv
langchain-openai
openai
fire

#metrics calculation
scikit-learn
pandas